content/canvas/test/test_2d.drawImage.nowrap.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.drawImage.nowrap</title>
<!-- Testing: Stretched images do not get pixels wrapping around the edges -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
function isPixel(ctx, x,y, r,g,b,a, pos, colour, d) {
    var pixel = ctx.getImageData(x, y, 1, 1);
    var pr = pixel.data[0],
        pg = pixel.data[1],
        pb = pixel.data[2],
        pa = pixel.data[3];
    ok(r-d <= pr && pr <= r+d &&
       g-d <= pg && pg <= g+d &&
       b-d <= pb && pb <= b+d &&
       a-d <= pa && pa <= a+d,
       "pixel "+pos+" is "+pr+","+pg+","+pb+","+pa+"; expected "+colour+" +/- "+d);
}

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);
ctx.drawImage(document.getElementById('redtransparent.png'), -1950, 0, 2000, 50);
isPixel(ctx, 45,25, 0,255,0,255, "45,25", "0,255,0,255", 2);
isPixel(ctx, 50,25, 0,255,0,255, "50,25", "0,255,0,255", 2);
isPixel(ctx, 55,25, 0,255,0,255, "55,25", "0,255,0,255", 2);

SimpleTest.finish();

});
</script>
<img src="image_redtransparent.png" id="redtransparent.png" class="resource">
